backend/requirements.txt
# --- API ---
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# --- DB ---
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1

# --- Config ---
pydantic-settings==2.2.1
python-dotenv==1.0.1

# --- Data / Excel ---
pandas==2.2.2
openpyxl==3.1.2

# --- Jobs async ---
redis==5.0.4
rq==1.16.2

# --- Utils ---
python-dateutil==2.9.0

# --- Auth ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3